requirements.txt
pathlib
matplotlib
pillow
natsort
tqdm
scikit-image
scikit-learn
tensorboard
tensorflow